% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{add_growth_form_taxa}
\alias{add_growth_form_taxa}
\title{Add growth forms to a single taxa}
\usage{
add_growth_form_taxa(idtax, con = NULL)
}
\value{
A tibble
}
\description{
Add growth form information to a single taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
