% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_subplot_features}
\alias{add_subplot_features}
\title{Add an observation in subplot_features table}
\usage{
add_subplot_features(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  plot_name_field = NULL,
  id_plot_name = NULL,
  id_plot_name_corresp = "id_table_liste_plots_n",
  subplottype_field,
  features_field = NULL,
  add_data = FALSE,
  ask_before_update = TRUE,
  interactive = TRUE,
  verbose = TRUE,
  check_existing_data = TRUE,
  con = NULL
)
}
\arguments{
\item{new_data}{tibble}

\item{col_names_select}{string vector}

\item{col_names_corresp}{string vector}

\item{plot_name_field}{string column name which contain the plot_name for linking}

\item{id_plot_name}{string id of plot_name}

\item{subplottype_field}{string vector listing trait columns names in new_data}

\item{add_data}{logical whether or not data should be added - by default FALSE}

\item{ask_before_update}{logical ask before adding}

\item{interactive}{logical whether to use interactive prompts (default TRUE)}

\item{verbose}{logical}

\item{check_existing_data}{logical if it should be checked if imported data already exist in the database}

\item{con}{database connection (optional, will create if NULL)}
}
\value{
list of tibbles that should be/have been added
}
\description{
Add a observation in subplot_features table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
