% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{add_subplottype}
\alias{add_subplottype}
\title{Add a type in subplot table}
\usage{
add_subplottype(
  new_type = NULL,
  new_valuetype = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_typedescription = NULL,
  new_factorlevels = NULL,
  new_expectedunit = NULL,
  new_comments = NULL,
  con = NULL,
  interactive = TRUE
)
}
\arguments{
\item{new_type}{string value with new type descriptors - try to avoid space}

\item{new_valuetype}{string one of following 'numeric', 'integer', 'categorical', 'ordinal', 'logical', 'character', 'table_colnam'}

\item{new_maxallowedvalue}{numeric if valuetype is numeric, indicate the maximum allowed value}

\item{new_minallowedvalue}{numeric if valuetype is numeric, indicate the minimum allowed value}

\item{new_typedescription}{string full description of type/feature}

\item{new_factorlevels}{string a vector of all possible value if valuetype is categorical or ordinal}

\item{new_expectedunit}{string expected unit (unitless if none)}

\item{new_comments}{string any comments}

\item{con}{Database connection (optional). If NULL, will create a new connection.}

\item{interactive}{Logical. If TRUE (default), will ask for confirmation. Set to FALSE when calling from Shiny.}
}
\value{
Invisibly returns the new feature data (tibble)
}
\description{
Add feature and associated descriptors in subplot list table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
