% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_tree_view.R
\name{build_breadcrumb_path}
\alias{build_breadcrumb_path}
\title{Build Breadcrumb Path HTML}
\usage{
build_breadcrumb_path(hierarchy, i18n)
}
\arguments{
\item{hierarchy}{Hierarchy list from get_taxon_hierarchy}

\item{i18n}{i18n translator}
}
\value{
Shiny tagList
}
\description{
Creates a breadcrumb-style path showing the hierarchy.
}
\keyword{internal}
