% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_tree_view.R
\name{build_hierarchy_tree_html}
\alias{build_hierarchy_tree_html}
\title{Build Hierarchy Tree HTML}
\usage{
build_hierarchy_tree_html(hierarchy, current_id, i18n)
}
\arguments{
\item{hierarchy}{Hierarchy list from get_taxon_hierarchy}

\item{current_id}{The ID of the currently selected taxon}

\item{i18n}{i18n translator}
}
\value{
Shiny tagList
}
\description{
Creates the nested tree view HTML structure.
}
\keyword{internal}
