% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{build_trait_query}
\alias{build_trait_query}
\title{Build appropriate SQL query based on parameters}
\usage{
build_trait_query(individual_ids, trait_ids, id_trait_measures = NULL, con)
}
\description{
Build appropriate SQL query based on parameters
}
\keyword{internal}
