% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_backbone.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if valid backbone cache exists}
\usage{
cache_exists()
}
\value{
Logical, TRUE if cache exists
}
\description{
Checks whether both the cache data file and metadata file exist.
}
\keyword{internal}
