% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_integration.R
\name{check_wcvp_update}
\alias{check_wcvp_update}
\title{Check if WCVP Update is Available}
\usage{
check_wcvp_update(con_taxa = NULL)
}
\arguments{
\item{con_taxa}{Connection to the taxa database. If NULL, calls \code{call.mydb.taxa()}.}
}
\value{
Logical. TRUE if a newer version is available.
}
\description{
Compares the database WCVP version with the version available in the
\code{rWCVP} package.
}
\examples{
\dontrun{
if (check_wcvp_update()) {
  import_wcvp_names(con_taxa, force = TRUE)
}
}

}
