% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step6_import.R
\name{.add_people_to_census}
\alias{.add_people_to_census}
\title{Add people features to newly created census records}
\usage{
.add_people_to_census(data, people_cols, con)
}
\description{
Add people features to newly created census records
}
\keyword{internal}
