% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_matching.R
\name{.add_synonym_info_r}
\alias{.add_synonym_info_r}
\title{Add synonym information to matches using cached backbone}
\usage{
.add_synonym_info_r(matches, backbone)
}
\description{
Add synonym information to matches using cached backbone
}
\keyword{internal}
