% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_individual_traits.R
\name{.aggregate_one_rule}
\alias{.aggregate_one_rule}
\title{Aggregate a single rule, return rows ready to insert into taxa_traits_measures}
\usage{
.aggregate_one_rule(
  rule,
  citation_id,
  con,
  con_taxa,
  allowed_tax_levels = c("species", "infraspecific")
)
}
\description{
Aggregate a single rule, return rows ready to insert into taxa_traits_measures
}
\keyword{internal}
