% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step4_lookup.R
\name{.apply_people_id_conversion}
\alias{.apply_people_id_conversion}
\title{Convert people name strings to database IDs in data}
\usage{
.apply_people_id_conversion(data, people_columns, name_to_id_map)
}
\description{
Convert people name strings to database IDs in data
}
\keyword{internal}
