% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_multi_stems.R
\name{.auto_detect_col}
\alias{.auto_detect_col}
\title{Auto-detect a column name from a list of candidates}
\usage{
.auto_detect_col(col_names, candidates)
}
\description{
Auto-detect a column name from a list of candidates
}
\keyword{internal}
