% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_measurements.R
\name{.auto_match_columns}
\alias{.auto_match_columns}
\title{Simple auto-matching of user columns to known column names via synonyms}
\usage{
.auto_match_columns(user_cols, synonyms)
}
\description{
Simple auto-matching of user columns to known column names via synonyms
}
\keyword{internal}
