% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{.build_combined_taxa_code}
\alias{.build_combined_taxa_code}
\title{Build the combined workflow script}
\usage{
.build_combined_taxa_code(
  query_taxa_code,
  traits_code,
  has_selection,
  is_public = FALSE
)
}
\description{
Build the combined workflow script
}
\keyword{internal}
