% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{.build_query_taxa_code}
\alias{.build_query_taxa_code}
\title{Build query_taxa() code string from search params}
\usage{
.build_query_taxa_code(params)
}
\description{
Build query_taxa() code string from search params
}
\keyword{internal}
