% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{.build_query_taxa_traits_code}
\alias{.build_query_taxa_traits_code}
\title{Build query_taxa_traits() code string from selected taxon IDs}
\usage{
.build_query_taxa_traits_code(idtax_ids)
}
\description{
Build query_taxa_traits() code string from selected taxon IDs
}
\keyword{internal}
