% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_census_import.R
\name{.census_mapped_columns}
\alias{.census_mapped_columns}
\title{Columns already claimed as keys or individual attributes}
\usage{
.census_mapped_columns(input)
}
\arguments{
\item{input}{Shiny input object.}
}
\value{
Character vector of column names.
}
\description{
Columns already claimed as keys or individual attributes
}
\keyword{internal}
