% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_census_import.R
\name{.census_safe_id}
\alias{.census_safe_id}
\title{Input id fragment for a user column name}
\usage{
.census_safe_id(col)
}
\arguments{
\item{col}{Character vector of column names.}
}
\value{
Character vector safe for use in an input id.
}
\description{
Two columns differing only in punctuation collapse onto the same id, which
is why the same rule has to be used everywhere a column is turned into an
input name.
}
\keyword{internal}
