% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_import_transaction.R
\name{.census_subplot_type_id}
\alias{.census_subplot_type_id}
\title{Resolve the `census` subplot type id}
\usage{
.census_subplot_type_id(con)
}
\arguments{
\item{con}{Database connection.}
}
\value{
Integer id from `subplotype_list`.
}
\description{
Resolve the `census` subplot type id
}
\keyword{internal}
