% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_census_import.R
\name{.census_trait_mapping}
\alias{.census_trait_mapping}
\title{Collect the column-to-trait mapping chosen in the UI}
\usage{
.census_trait_mapping(input, data, traits)
}
\arguments{
\item{input}{Shiny input object.}

\item{data}{Data frame whose columns were offered.}

\item{traits}{Trait table from `traitlist`.}
}
\value{
Named character vector, column name to trait name.
}
\description{
Collect the column-to-trait mapping chosen in the UI
}
\keyword{internal}
