% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_standardization.R
\name{.classify_phrases}
\alias{.classify_phrases}
\title{Classify phrases against the ontology}
\usage{
.classify_phrases(phrases, onto)
}
\description{
Classify phrases against the ontology
}
\keyword{internal}
