% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_multi_stems.R
\name{.compute_stem_order}
\alias{.compute_stem_order}
\title{Compute stem_order within each group}
\usage{
.compute_stem_order(data)
}
\description{
Compute stem_order within each group
}
\keyword{internal}
