% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{.copy_btn}
\alias{.copy_btn}
\title{Render a rclipboard copy button}
\usage{
.copy_btn(ns, btn_id, code_str, i18n, label = NULL)
}
\description{
Render a rclipboard copy button
}
\keyword{internal}
