% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.create_hierarchy_entry_for_parent}
\alias{.create_hierarchy_entry_for_parent}
\title{Create a Hierarchy Entry for Parent (internal helper)}
\usage{
.create_hierarchy_entry_for_parent(
  con,
  tax_gen = NA,
  tax_esp = NA,
  tax_fam = NA,
  tax_order = NA,
  tax_famclass = NA,
  tax_level = NA,
  id_parent = NA
)
}
\arguments{
\item{con}{Database connection}

\item{tax_gen}{Genus name}

\item{tax_esp}{Species epithet}

\item{tax_fam}{Family name}

\item{tax_order}{Order name}

\item{tax_famclass}{Class name}

\item{tax_level}{Taxonomic level: "class", "order", "family", "genus", "species"}

\item{id_parent}{Parent taxon ID}
}
\description{
Creates a new entry at the specified taxonomic level with proper id_parent.
Uses tax_level column to indicate the taxonomic rank.
}
\keyword{internal}
