% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{.create_matching_row}
\alias{.create_matching_row}
\title{Create Matching Row for One Value}
\usage{
.create_matching_row(
  user_value,
  column_name,
  lookup_info,
  session,
  suggestions = NULL
)
}
\description{
Create Matching Row for One Value
}
\keyword{internal}
