% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{.dark_code_block}
\alias{.dark_code_block}
\title{Render a dark-background code block}
\usage{
.dark_code_block(ns, output_id, code_str)
}
\description{
Render a dark-background code block
}
\keyword{internal}
