% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_new_plot_processing.R
\name{.decode_code_column}
\alias{.decode_code_column}
\title{Decode a coded column against an OpenForis code list}
\usage{
.decode_code_column(values, code_list, code_col, label_col)
}
\arguments{
\item{values}{Vector of codes (numeric or character).}

\item{code_list}{Parsed code list data frame, or NULL.}

\item{code_col}{Name of the code column in \code{code_list}.}

\item{label_col}{Name of the label column in \code{code_list}.}
}
\value{
Character vector of labels, NA where decoding failed.
}
\description{
Decode a coded column against an OpenForis code list
}
\keyword{internal}
