% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.decode_openforis_light}
\alias{.decode_openforis_light}
\title{Decode OpenForis light column}
\usage{
.decode_openforis_light(data, code_list)
}
\description{
Decode OpenForis light column
}
\keyword{internal}
