% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.decode_openforis_pom}
\alias{.decode_openforis_pom}
\title{Decode OpenForis POM observation columns}
\usage{
.decode_openforis_pom(data, code_list)
}
\description{
Decode OpenForis POM observation columns
}
\keyword{internal}
