% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.default_observation_flags}
\alias{.default_observation_flags}
\title{Default observation flag mapping}
\usage{
.default_observation_flags()
}
\description{
Default observation flag mapping
}
\keyword{internal}
