% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.default_status_flags}
\alias{.default_status_flags}
\title{Default stem status flag mapping}
\usage{
.default_status_flags()
}
\description{
Default stem status flag mapping
}
\keyword{internal}
