% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_country}
\alias{.delete_country}
\title{Delete an entry in country table}
\usage{
.delete_country(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in country table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
