% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step6_import.R
\name{.execute_feature_import}
\alias{.execute_feature_import}
\title{Execute feature import (internal)}
\usage{
.execute_feature_import(data, config, con, dry_run = TRUE, i18n = NULL)
}
\arguments{
\item{data}{Data frame of features to import}

\item{config}{Feature configuration list}

\item{con}{Database connection}

\item{dry_run}{Logical, if TRUE only simulate}

\item{i18n}{Translator object}
}
\value{
List with import results
}
\description{
Execute feature import (internal)
}
\keyword{internal}
