% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.extract_linked_individuals_from_specimens}
\alias{.extract_linked_individuals_from_specimens}
\title{Extract Linked Individuals from Specimens}
\usage{
.extract_linked_individuals_from_specimens(specimen_ids, con)
}
\arguments{
\item{specimen_ids}{Vector of specimen IDs}

\item{con}{Database connection}
}
\value{
Data frame with linked individuals
}
\description{
Internal helper to get individuals linked to specimens
}
\keyword{internal}
