% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.extract_numeric_traits}
\alias{.extract_numeric_traits}
\title{Extract direct numeric measurements into long format}
\usage{
.extract_numeric_traits(data, numeric_cols)
}
\description{
Extract direct numeric measurements into long format
}
\keyword{internal}
