% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_overview.R
\name{.feature_rule_labels}
\alias{.feature_rule_labels}
\title{`.feature_rule_label()` over vectors}
\usage{
.feature_rule_labels(rules, ns_records, i18n)
}
\description{
`.feature_rule_label()` over vectors
}
\keyword{internal}
