% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{.get_fuzzy_matches}
\alias{.get_fuzzy_matches}
\title{Get Fuzzy Matches for a Value}
\usage{
.get_fuzzy_matches(user_value, lookup_info, n = NULL)
}
\description{
Get Fuzzy Matches for a Value
}
\keyword{internal}
