% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{.get_lookup_info}
\alias{.get_lookup_info}
\title{Get Lookup Table Info}
\usage{
.get_lookup_info(column_name, con, extra_people_cols = NULL)
}
\description{
Get Lookup Table Info
}
\keyword{internal}
