% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_features_user_friendly.R
\name{.get_subplot_feature_synonyms}
\alias{.get_subplot_feature_synonyms}
\title{Get Subplot Feature Synonyms (Internal Helper)}
\usage{
.get_subplot_feature_synonyms()
}
\description{
Returns common synonyms for subplot feature types.
}
\keyword{internal}
