% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot_features_user_friendly.R
\name{.identify_plot_id_column}
\alias{.identify_plot_id_column}
\title{Identify Plot ID Column (Internal Helper)}
\usage{
.identify_plot_id_column(data, plot_id_column, interactive, verbose)
}
\description{
Identify Plot ID Column (Internal Helper)
}
\keyword{internal}
