% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.include_children}
\alias{.include_children}
\title{Recursively include all child taxa via id_parent}
\usage{
.include_children(res, mydb_taxa, verbose)
}
\arguments{
\item{res}{data.frame of initial taxa results}

\item{mydb_taxa}{database connection}

\item{verbose}{logical}
}
\value{
data.frame with children appended
}
\description{
Recursively include all child taxa via id_parent
}
\keyword{internal}
