% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_import_transaction.R
\name{.insert_census_recruits}
\alias{.insert_census_recruits}
\title{Insert recruits into `data_individuals`}
\usage{
.insert_census_recruits(recruits, con)
}
\arguments{
\item{recruits}{Data frame with `id_table_liste_plots_n`, `tag` and any of
the optional individual columns.}

\item{con}{Database connection.}
}
\value{
Data frame with `id_individuals`, `tag` and `plot_name`.
}
\description{
Insert recruits into `data_individuals`
}
\keyword{internal}
