% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_xlsx_upload.R
\name{.is_excel_file}
\alias{.is_excel_file}
\title{Is an uploaded file an Excel workbook?}
\usage{
.is_excel_file(name)
}
\arguments{
\item{name}{File name (with extension).}
}
\value{
Logical scalar.
}
\description{
Is an uploaded file an Excel workbook?
}
\keyword{internal}
