% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.link_subplotype}
\alias{.link_subplotype}
\title{Internal function}
\usage{
.link_subplotype(data_stand, subplotype, con = NULL)
}
\arguments{
\item{data_stand}{tibble}

\item{subplotype}{string vector}

\item{con}{Database connection (required)}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
