% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{.make_input_id}
\alias{.make_input_id}
\title{Make Input ID from Value and Column}
\usage{
.make_input_id(user_value, column_name)
}
\description{
Make Input ID from Value and Column
}
\keyword{internal}
