% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.map_status_flags}
\alias{.map_status_flags}
\title{Map stem status labels to flag codes}
\usage{
.map_status_flags(label1, label2, type = "alive", flag_mapping)
}
\description{
Map stem status labels to flag codes
}
\keyword{internal}
