% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_matching.R
\name{.match_exact_r}
\alias{.match_exact_r}
\title{Exact match against cached backbone}
\usage{
.match_exact_r(parsed, backbone, include_authors, max_matches)
}
\description{
Exact match against cached backbone
}
\keyword{internal}
