% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_lookup.R
\name{.match_lookup_values}
\alias{.match_lookup_values}
\title{Match Lookup Values (Internal Helper)}
\usage{
.match_lookup_values(user_values, db_values, db_ids)
}
\arguments{
\item{user_values}{Vector of user-provided values}

\item{db_values}{Vector of database values}

\item{db_ids}{Vector of database IDs}
}
\value{
List with exact matches, fuzzy matches, and unmatched values
}
\description{
Match Lookup Values (Internal Helper)
}
\keyword{internal}
