% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_multi_stems.R
\name{.multi_stems_upload_ui}
\alias{.multi_stems_upload_ui}
\title{Upload sub-UI for multi-stems}
\usage{
.multi_stems_upload_ui(ns, i18n)
}
\description{
Upload sub-UI for multi-stems
}
\keyword{internal}
