% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openforis_processing.R
\name{.pivot_coded_columns}
\alias{.pivot_coded_columns}
\title{Pivot coded columns to long format and decode via code list}
\usage{
.pivot_coded_columns(
  data,
  coded_cols,
  code_list,
  code_col = "item_code",
  label_col = "item_label_en"
)
}
\description{
Pivot coded columns to long format and decode via code list
}
\keyword{internal}
