% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen_linking_functions.R
\name{.query_unmatched_specimens}
\alias{.query_unmatched_specimens}
\title{Query Unmatched Specimens (Internal)}
\usage{
.query_unmatched_specimens(con = NULL)
}
\arguments{
\item{con}{Database connection. If NULL, calls call.mydb()}
}
\value{
List with problematic specimens
}
\description{
Find specimens that have taxonomic discrepancies or are not properly linked.
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
