% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_measurements.R
\name{.render_long_trait_mapping_ui}
\alias{.render_long_trait_mapping_ui}
\title{Render trait mapping UI for long format}
\usage{
.render_long_trait_mapping_ui(ns, i18n, raw, traits, input)
}
\description{
Render trait mapping UI for long format
}
\keyword{internal}
