% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_measurements.R
\name{.render_wide_mapping_ui}
\alias{.render_wide_mapping_ui}
\title{Render column mapping UI for wide format}
\usage{
.render_wide_mapping_ui(ns, i18n, user_cols, none, auto_map)
}
\description{
Render column mapping UI for wide format
}
\keyword{internal}
