% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxo_match_r_code.R
\name{.taxo_match_col}
\alias{.taxo_match_col}
\title{Quote a column name for use inside generated `[[ ]]` indexing}
\usage{
.taxo_match_col(x)
}
\description{
Quote a column name for use inside generated `[[ ]]` indexing
}
\keyword{internal}
