% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxo_match_r_code.R
\name{.taxo_match_combined_code}
\alias{.taxo_match_combined_code}
\title{Assemble the full workflow script}
\usage{
.taxo_match_combined_code(
  prep_code,
  match_code,
  backbone_code,
  is_offline = FALSE
)
}
\description{
Assemble the full workflow script
}
\keyword{internal}
