% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxo_match_r_code.R
\name{.taxo_match_matching_code}
\alias{.taxo_match_matching_code}
\title{Build the `standardize_taxonomic_batch()` code}
\usage{
.taxo_match_matching_code(params)
}
\arguments{
\item{params}{Settings captured at the last matching run}
}
\description{
Build the `standardize_taxonomic_batch()` code
}
\keyword{internal}
