% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxo_match_r_code.R
\name{.taxo_match_wcvp_code}
\alias{.taxo_match_wcvp_code}
\title{Build the optional WCVP enrichment code}
\usage{
.taxo_match_wcvp_code()
}
\description{
Build the optional WCVP enrichment code
}
\keyword{internal}
