% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_column_types}
\alias{.upd_column_types}
\title{Column types of a table, as a named character vector}
\usage{
.upd_column_types(table, con)
}
\description{
Column types of a table, as a named character vector
}
\keyword{internal}
