% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_display_value}
\alias{.upd_display_value}
\title{Human-readable value of a feature record}
\usage{
.upd_display_value(valuetype, value_num, value_char, lookup_id, colnam_lookup)
}
\description{
Human-readable value of a feature record
}
\keyword{internal}
