% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_entity_spec}
\alias{.upd_entity_spec}
\title{Table / id-column / update_records table_type for an entity}
\usage{
.upd_entity_spec(entity = c("plot", "individual"))
}
\description{
Table / id-column / update_records table_type for an entity
}
\keyword{internal}
