% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_fetch_record}
\alias{.upd_fetch_record}
\title{Fetch the current row of a plot or an individual}
\usage{
.upd_fetch_record(entity, id, con)
}
\description{
Fetch the current row of a plot or an individual
}
\keyword{internal}
