% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_record.R
\name{.upd_link_problem_text}
\alias{.upd_link_problem_text}
\title{A parent-link problem code as a sentence}
\usage{
.upd_link_problem_text(code, i18n)
}
\arguments{
\item{code}{One code from [.upd_validate_plot_link()].}

\item{i18n}{A resolved `shiny.i18n` translator.}
}
\value{
A single string.
}
\description{
A parent-link problem code as a sentence
}
\keyword{internal}
