% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_lookup_choices}
\alias{.upd_lookup_choices}
\title{Choices for a lookup column, as a named vector (label -> id)}
\usage{
.upd_lookup_choices(lookup_table, lookup_key, lookup_value, con)
}
\description{
Choices for a lookup column, as a named vector (label -> id)
}
\keyword{internal}
