% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_record_dates}
\alias{.upd_record_dates}
\title{Dates of feature records as `YYYY`, `YYYY-MM` or `YYYY-MM-DD`}
\usage{
.upd_record_dates(grp)
}
\description{
Dates of feature records as `YYYY`, `YYYY-MM` or `YYYY-MM-DD`
}
\keyword{internal}
