% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_app_resolver.R
\name{.upd_specimen_label}
\alias{.upd_specimen_label}
\title{A collector's label for a specimen: "Dauby 1234b"}
\usage{
.upd_specimen_label(specimen)
}
\description{
A collector's label for a specimen: "Dauby 1234b"
}
\keyword{internal}
