% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_feature_wizard.R
\name{feature_wizard_server}
\alias{feature_wizard_server}
\title{Server for Feature Wizard}
\usage{
feature_wizard_server(input, output, session, translator)
}
\description{
Server for Feature Wizard
}
\keyword{internal}
