% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_feature_wizard.R
\name{feature_wizard_ui}
\alias{feature_wizard_ui}
\title{UI for Feature Wizard}
\usage{
feature_wizard_ui(translator, language = "fr")
}
\description{
UI for Feature Wizard
}
\keyword{internal}
