% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{fetch_measurement_features_raw}
\alias{fetch_measurement_features_raw}
\title{Fetch raw measurement features data}
\usage{
fetch_measurement_features_raw(id_trait_measures, id_trait = NULL, config, con)
}
\description{
Fetch raw measurement features data
}
\keyword{internal}
