% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{fetch_trait_measurements}
\alias{fetch_trait_measurements}
\title{Fetch trait measurements with automatic chunking}
\usage{
fetch_trait_measurements(
  individual_ids,
  trait_ids,
  id_trait_measures = NULL,
  con
)
}
\description{
Fetch trait measurements with automatic chunking
}
\keyword{internal}
