% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{filter_to_census}
\alias{filter_to_census}
\title{Filter data to first or last census}
\usage{
filter_to_census(data, strategy = c("first", "last"))
}
\description{
Filter data to first or last census
}
\keyword{internal}
