% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_consistency.R
\name{fix_hierarchy_inconsistencies}
\alias{fix_hierarchy_inconsistencies}
\title{Fix Hierarchy Inconsistencies}
\usage{
fix_hierarchy_inconsistencies(con, inconsistencies)
}
\arguments{
\item{con}{Database connection}

\item{inconsistencies}{List of inconsistencies from check_hierarchy_consistency}
}
\value{
Summary of fixes applied
}
\description{
Updates flat columns to match the hierarchy defined by id_parent.
}
\keyword{internal}
