% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_i18n.R
\name{get_available_languages}
\alias{get_available_languages}
\title{Get Available Translation Languages}
\usage{
get_available_languages(translator)
}
\arguments{
\item{translator}{shiny.i18n::Translator object}
}
\value{
Character vector of language codes
}
\description{
Helper to get list of available languages from translator
}
\keyword{internal}
