% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{get_taxon_ancestors}
\alias{get_taxon_ancestors}
\title{Get All Ancestors of a Taxon}
\usage{
get_taxon_ancestors(idtax_n, con = NULL, include_self = TRUE)
}
\arguments{
\item{idtax_n}{The taxon ID to find ancestors for}

\item{con}{Database connection (optional, will create if NULL)}

\item{include_self}{If TRUE, include the taxon itself in results}
}
\value{
Data frame of ancestor taxa, ordered from root to taxon
}
\description{
Walks up the id_parent chain to return all ancestor taxa from the given
taxon up to the root (class level).
}
