% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_update_functions_old.R
\name{get_updates_diconame}
\alias{get_updates_diconame}
\title{Get backups of modified taxonomic data}
\usage{
get_updates_diconame(
  id = NULL,
  last_months = NULL,
  last_10_entry = TRUE,
  last = NULL
)
}
\arguments{
\item{id}{look backups for a specific id (of taxonomic table)}

\item{last_months}{look backups performed this last month}

\item{last_10_entry}{look the last 10 backups performed}

\item{last}{look the last backup}
}
\value{
A tibble
}
\description{
List taxonomic data that has been modified
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
