% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_import_wizard.R
\name{import_wizard_server}
\alias{import_wizard_server}
\title{Server for Import Wizard}
\usage{
import_wizard_server(input, output, session, translator)
}
\description{
Server for Import Wizard
}
\keyword{internal}
