% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_i18n.R
\name{init_translator}
\alias{init_translator}
\title{Initialize Translator for Shiny App}
\usage{
init_translator()
}
\value{
shiny.i18n::Translator object
}
\description{
Creates a shiny.i18n Translator object loaded with translations from the JSON file.
This should be called once at app initialization.
}
\details{
The translator reads from `inst/translations/translation.json` which contains
translations for English (en) and French (fr).
}
\examples{
\dontrun{
translator <- init_translator()
translator$set_translation_language("fr")
translator$t("app_title")
}

}
\keyword{internal}
