% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{match_tax}
\alias{match_tax}
\title{Query and standardize taxonomy}
\usage{
match_tax(idtax, queried_tax = NULL, verbose = TRUE, backbone = "internal")
}
\arguments{
\item{idtax}{vector of idtax_n to be search}

\item{queried_tax}{tibble, output of query_taxa}

\item{verbose}{logical whether results should be shown in viewer}

\item{backbone}{Character. Backbone whose names are used: \code{"internal"}
(default) for \code{table_taxa}, or the code of a backbone registered in
the taxa database (see \code{list_backbones()}), such as \code{"wcvp"}.}
}
\value{
tibble
}
\description{
Query and standardize taxonomy for synonymies and add traits information at species and genus levels
}
\examples{
\dontrun{
match_tax(idtax = c(3095, 219))
}

}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
