% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_backbone_cache_selection.R
\name{mod_backbone_cache_selection_ui}
\alias{mod_backbone_cache_selection_ui}
\title{Backbone Cache Selection Module - UI}
\usage{
mod_backbone_cache_selection_ui(id)
}
\arguments{
\item{id}{Character, module namespace ID}
}
\value{
Shiny UI element
}
\description{
UI component for backbone cache selection modal.
The modal is created dynamically in the server, so this returns an empty div.
}
\keyword{internal}
