% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_census_information.R
\name{mod_census_information_server}
\alias{mod_census_information_server}
\title{Census Information Module - Server}
\usage{
mod_census_information_server(id, imported_plots, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{imported_plots}{Reactive containing the imported plot data with plot_name and id_liste_plots}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object from shiny.i18n}
}
\value{
Reactive containing census addition result
}
\description{
Census Information Module - Server
}
\keyword{internal}
