% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_census_information.R
\name{mod_census_information_ui}
\alias{mod_census_information_ui}
\title{Census Information Module - UI}
\usage{
mod_census_information_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Census Information Module - UI
}
\keyword{internal}
