% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_citation_panel.R
\name{mod_citation_panel_ui}
\alias{mod_citation_panel_ui}
\title{Citation Panel Module - UI}
\usage{
mod_citation_panel_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
shiny tagList
}
\description{
Citation Panel Module - UI
}
\keyword{internal}
