% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_code_preview.R
\name{mod_code_preview_server}
\alias{mod_code_preview_server}
\title{Code Preview Module - Server}
\usage{
mod_code_preview_server(
  id,
  filters,
  selected_plots,
  extraction_options,
  metadata_available,
  individuals_available,
  i18n,
  individual_features_options = NULL,
  individual_features_available = NULL
)
}
\arguments{
\item{id}{Module namespace ID}

\item{filters}{Reactive returning named list of filter values}

\item{selected_plots}{Reactive returning vector of selected plot IDs}

\item{extraction_options}{Reactive returning named list of extraction options}

\item{metadata_available}{Reactive returning TRUE when metadata has been queried}

\item{individuals_available}{Reactive returning TRUE when individuals have been extracted}

\item{i18n}{Reactive returning shiny.i18n translator}

\item{individual_features_options}{Reactive returning named list of individual features options (optional)}

\item{individual_features_available}{Reactive returning TRUE when individual features have been queried (optional)}
}
\description{
Server logic for generating and displaying equivalent R code
}
\keyword{internal}
