% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_code_preview.R
\name{mod_code_preview_ui}
\alias{mod_code_preview_ui}
\title{Code Preview Module - UI}
\usage{
mod_code_preview_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying equivalent R code
}
\keyword{internal}
