% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_column_select.R
\name{mod_column_select_server}
\alias{mod_column_select_server}
\title{Column Select Module - Server}
\usage{
mod_column_select_server(id, data, initial_column = NULL, i18n)
}
\arguments{
\item{id}{Character, module ID}

\item{data}{Reactive data.frame from data input module}

\item{initial_column}{Character, optional pre-selected column name}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive list with $column (selected column name), $include_authors
  (logical), $data (potentially modified data), $mode ("single" or
  "multiple") and, in multiple mode, $genus_column / $species_column /
  $family_column
}
\description{
Column Select Module - Server
}
\keyword{internal}
