% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_server}
\alias{mod_data_input_server}
\title{Data Input Module - Server}
\usage{
mod_data_input_server(id, provided_data = NULL, i18n)
}
\arguments{
\item{id}{Character, module ID}

\item{provided_data}{Reactive or data.frame, optional pre-loaded data}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive data.frame with user data
}
\description{
Data Input Module - Server
}
\keyword{internal}
