% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extraction_config.R
\name{mod_extraction_config_ui}
\alias{mod_extraction_config_ui}
\title{Extraction Configuration Module - UI}
\usage{
mod_extraction_config_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for configuring individual extraction options
}
\keyword{internal}
