% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step1_select_plots.R
\name{mod_feat_step1_select_plots_server}
\alias{mod_feat_step1_select_plots_server}
\title{Feature Wizard Step 1: Select Plots - Server}
\usage{
mod_feat_step1_select_plots_server(id, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing data.frame of selected plots
}
\description{
Feature Wizard Step 1: Select Plots - Server
}
\keyword{internal}
