% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step1_select_plots.R
\name{mod_feat_step1_select_plots_ui}
\alias{mod_feat_step1_select_plots_ui}
\title{Feature Wizard Step 1: Select Plots - UI}
\usage{
mod_feat_step1_select_plots_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 1: Select Plots - UI
}
\keyword{internal}
