% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step2_choose_mode.R
\name{mod_feat_step2_choose_mode_server}
\alias{mod_feat_step2_choose_mode_server}
\title{Feature Wizard Step 2: Choose Mode - Server}
\usage{
mod_feat_step2_choose_mode_server(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing selected mode string
}
\description{
Feature Wizard Step 2: Choose Mode - Server
}
\keyword{internal}
