% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step2_choose_mode.R
\name{mod_feat_step2_choose_mode_ui}
\alias{mod_feat_step2_choose_mode_ui}
\title{Feature Wizard Step 2: Choose Mode - UI}
\usage{
mod_feat_step2_choose_mode_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 2: Choose Mode - UI
}
\keyword{internal}
