% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_census_import.R
\name{mod_feat_step3_census_import_server}
\alias{mod_feat_step3_census_import_server}
\title{Feature Wizard Step 3: Full Census Import - Server}
\usage{
mod_feat_step3_census_import_server(id, selected_plots, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{selected_plots}{Reactive containing data.frame of selected plots}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing list(data, config)
}
\description{
Feature Wizard Step 3: Full Census Import - Server
}
\keyword{internal}
