% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_census_import.R
\name{mod_feat_step3_census_import_ui}
\alias{mod_feat_step3_census_import_ui}
\title{Feature Wizard Step 3: Full Census Import - UI}
\usage{
mod_feat_step3_census_import_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 3: Full Census Import - UI
}
\keyword{internal}
