% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_multi_stems.R
\name{mod_feat_step3_multi_stems_server}
\alias{mod_feat_step3_multi_stems_server}
\title{Feature Wizard Step 3: Multi-Stems - Server}
\usage{
mod_feat_step3_multi_stems_server(
  id,
  selected_plots,
  operation_mode,
  con,
  i18n
)
}
\arguments{
\item{id}{Module namespace ID}

\item{selected_plots}{Reactive containing selected plots data}

\item{operation_mode}{Reactive containing operation mode string}

\item{con}{Reactive returning database connection/pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing list(data, config) for downstream steps
}
\description{
Feature Wizard Step 3: Multi-Stems - Server
}
\keyword{internal}
