% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_multi_stems.R
\name{mod_feat_step3_multi_stems_ui}
\alias{mod_feat_step3_multi_stems_ui}
\title{Feature Wizard Step 3: Multi-Stems - UI}
\usage{
mod_feat_step3_multi_stems_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 3: Multi-Stems - UI
}
\keyword{internal}
