% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_plot_features.R
\name{mod_feat_step3_plot_features_server}
\alias{mod_feat_step3_plot_features_server}
\title{Feature Wizard Step 3: Plot Features - Server}
\usage{
mod_feat_step3_plot_features_server(
  id,
  selected_plots,
  operation_mode,
  con,
  i18n
)
}
\arguments{
\item{id}{Module namespace ID}

\item{selected_plots}{Reactive containing data.frame of selected plots}

\item{operation_mode}{Reactive containing mode string}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing list(data, config)
}
\description{
Feature Wizard Step 3: Plot Features - Server
}
\keyword{internal}
