% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_stem_status.R
\name{mod_feat_step3_stem_status_ui}
\alias{mod_feat_step3_stem_status_ui}
\title{Feature Wizard Step 3: Compute Stem Status - UI}
\usage{
mod_feat_step3_stem_status_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 3: Compute Stem Status - UI
}
\keyword{internal}
