% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3b_taxon_revision.R
\name{mod_feat_step3b_taxon_revision_ui}
\alias{mod_feat_step3b_taxon_revision_ui}
\title{UI for the taxon revision step}
\usage{
mod_feat_step3b_taxon_revision_ui(id, i18n)
}
\arguments{
\item{id}{Module id.}

\item{i18n}{Translator object.}
}
\value{
A `shiny::tagList`.
}
\description{
UI for the taxon revision step
}
\keyword{internal}
