% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step4_lookup.R
\name{mod_feat_step4_lookup_server}
\alias{mod_feat_step4_lookup_server}
\title{Feature Wizard Step 4: Lookup Matching - Server}
\usage{
mod_feat_step4_lookup_server(id, feature_data, feature_config, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{feature_data}{Reactive containing the prepared feature data}

\item{feature_config}{Reactive containing the feature configuration}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
List of reactives: data (matched data), complete (boolean)
}
\description{
Feature Wizard Step 4: Lookup Matching - Server
}
\keyword{internal}
