% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step4_lookup.R
\name{mod_feat_step4_lookup_ui}
\alias{mod_feat_step4_lookup_ui}
\title{Feature Wizard Step 4: Lookup Matching - UI}
\usage{
mod_feat_step4_lookup_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 4: Lookup Matching - UI
}
\keyword{internal}
