% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step5_validation.R
\name{mod_feat_step5_validation_server}
\alias{mod_feat_step5_validation_server}
\title{Feature Wizard Step 5: Validation - Server}
\usage{
mod_feat_step5_validation_server(
  id,
  matched_data,
  feature_config,
  selected_plots,
  operation_mode,
  con,
  i18n
)
}
\arguments{
\item{id}{Module namespace ID}

\item{matched_data}{Reactive containing the matched feature data}

\item{feature_config}{Reactive containing the feature configuration}

\item{selected_plots}{Reactive containing selected plots data}

\item{operation_mode}{Reactive containing operation mode string}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive containing validation result
}
\description{
Feature Wizard Step 5: Validation - Server
}
\keyword{internal}
