% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step5_validation.R
\name{mod_feat_step5_validation_ui}
\alias{mod_feat_step5_validation_ui}
\title{Feature Wizard Step 5: Validation - UI}
\usage{
mod_feat_step5_validation_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Feature Wizard Step 5: Validation - UI
}
\keyword{internal}
