% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_growth_form_selector.R
\name{mod_growth_form_selector_ui}
\alias{mod_growth_form_selector_ui}
\title{Growth Form Selector Module - UI}
\usage{
mod_growth_form_selector_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for selecting growth forms hierarchically
}
\keyword{internal}
