% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_herbarium_parser.R
\name{mod_herbarium_parser_server}
\alias{mod_herbarium_parser_server}
\title{Herbarium Parser Module - Server}
\usage{
mod_herbarium_parser_server(id, individuals_data, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{individuals_data}{Reactive containing individuals data from Step 1}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing parsed data
}
\description{
Herbarium Parser Module - Server
}
\keyword{internal}
