% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_herbarium_parser.R
\name{mod_herbarium_parser_ui}
\alias{mod_herbarium_parser_ui}
\title{Herbarium Parser Module - UI}
\usage{
mod_herbarium_parser_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Herbarium Parser Module - UI
}
\keyword{internal}
