% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_link_executor.R
\name{mod_link_executor_server}
\alias{mod_link_executor_server}
\title{Link Executor Module - Server}
\usage{
mod_link_executor_server(id, validated_links, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{validated_links}{Reactive containing validated links from Step 5}

\item{con}{Reactive database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing execution results
}
\description{
Link Executor Module - Server
}
\keyword{internal}
