% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_link_executor.R
\name{mod_link_executor_ui}
\alias{mod_link_executor_ui}
\title{Link Executor Module - UI}
\usage{
mod_link_executor_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Link Executor Module - UI
}
\keyword{internal}
