% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_link_preview.R
\name{mod_link_preview_server}
\alias{mod_link_preview_server}
\title{Link Preview Module - Server}
\usage{
mod_link_preview_server(
  id,
  pool_main,
  selected_specimens,
  selected_individuals,
  i18n
)
}
\arguments{
\item{id}{Character, module namespace ID}

\item{pool_main}{Reactive returning database connection pool}

\item{selected_specimens}{Reactive returning selected specimens data}

\item{selected_individuals}{Reactive returning selected individuals data}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive with link creation results
}
\description{
Link Preview Module - Server
}
