% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_link_preview.R
\name{mod_link_preview_ui}
\alias{mod_link_preview_ui}
\title{Link Preview Module - UI}
\usage{
mod_link_preview_ui(id, i18n)
}
\arguments{
\item{id}{Character, module namespace ID}

\item{i18n}{shiny.i18n translator object}
}
\value{
Shiny UI element
}
\description{
Link Preview Module - UI
}
