% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{mod_lookup_matcher_ui}
\alias{mod_lookup_matcher_ui}
\title{Lookup Matcher Module - UI}
\usage{
mod_lookup_matcher_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Lookup Matcher Module - UI
}
\keyword{internal}
