% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_name_preview.R
\name{mod_name_preview_ui}
\alias{mod_name_preview_ui}
\title{Name Preview Module - UI}
\usage{
mod_name_preview_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Name Preview Module - UI
}
\keyword{internal}
