% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_filters.R
\name{mod_plot_filters_server}
\alias{mod_plot_filters_server}
\title{Plot Filters Module - Server}
\usage{
mod_plot_filters_server(id, pool, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{pool}{Database connection pool (reactive or static)}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
A reactive list containing:
  - filters: Named list of filter values
  - execute_trigger: Reactive counter that increments on execute
}
\description{
Server logic for plot query filters
}
\keyword{internal}
