% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_filters.R
\name{mod_plot_filters_ui}
\alias{mod_plot_filters_ui}
\title{Plot Filters Module - UI}
\usage{
mod_plot_filters_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for plot query filters
}
\keyword{internal}
