% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_metadata_viewer.R
\name{mod_plot_metadata_viewer_ui}
\alias{mod_plot_metadata_viewer_ui}
\title{Plot Metadata Viewer Module - UI}
\usage{
mod_plot_metadata_viewer_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying plot metadata with interactive map and table
}
\keyword{internal}
