% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_statistics.R
\name{mod_plot_statistics_server}
\alias{mod_plot_statistics_server}
\title{Plot Statistics Module - Server}
\usage{
mod_plot_statistics_server(id, results, i18n)
}
\arguments{
\item{id}{Namespace ID}

\item{results}{Reactive containing query results (individuals data)}

\item{i18n}{Reactive translator object}
}
\value{
None (module handles its own outputs)
}
\description{
Server logic for plot statistics visualization
}
\keyword{internal}
