% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_display.R
\name{mod_results_display_ui}
\alias{mod_results_display_ui}
\title{Results Display Module - UI}
\usage{
mod_results_display_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying query results and download options
}
\keyword{internal}
