% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_lookup.R
\name{mod_specid_batch_lookup_server}
\alias{mod_specid_batch_lookup_server}
\title{Batch lookup server}
\usage{
mod_specid_batch_lookup_server(id, data, mappings, pool_main, i18n)
}
\arguments{
\item{id}{Module id}

\item{data}{Reactive uploaded data}

\item{mappings}{Reactive mappings list}

\item{pool_main}{Reactive main DB pool}

\item{i18n}{Reactive translator}
}
\value{
list(matched_data, is_complete)
}
\description{
Batch lookup server
}
\keyword{internal}
