% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_lookup.R
\name{mod_specid_batch_lookup_ui}
\alias{mod_specid_batch_lookup_ui}
\title{Batch lookup UI}
\usage{
mod_specid_batch_lookup_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Batch lookup UI
}
\keyword{internal}
