% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_mapping.R
\name{mod_specid_batch_mapping_server}
\alias{mod_specid_batch_mapping_server}
\title{Batch mapping server}
\usage{
mod_specid_batch_mapping_server(id, data, i18n)
}
\arguments{
\item{id}{Module id}

\item{data}{Reactive uploaded data frame}

\item{i18n}{Reactive translator}
}
\value{
list(mappings = reactive(named list), is_valid = reactive(logical))
}
\description{
Batch mapping server
}
\keyword{internal}
