% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_mapping.R
\name{mod_specid_batch_mapping_ui}
\alias{mod_specid_batch_mapping_ui}
\title{Batch mapping UI}
\usage{
mod_specid_batch_mapping_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Batch mapping UI
}
\keyword{internal}
