% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_update.R
\name{mod_specid_batch_update_server}
\alias{mod_specid_batch_update_server}
\title{Batch update server}
\usage{
mod_specid_batch_update_server(id, validated_data, mappings, pool_main, i18n)
}
\arguments{
\item{id}{Module id}

\item{validated_data}{Reactive validated rows}

\item{mappings}{Reactive mappings list}

\item{pool_main}{Reactive main DB pool}

\item{i18n}{Reactive translator}
}
\description{
Batch update server
}
\keyword{internal}
