% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_update.R
\name{mod_specid_batch_update_ui}
\alias{mod_specid_batch_update_ui}
\title{Batch update UI}
\usage{
mod_specid_batch_update_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Batch update UI
}
\keyword{internal}
