% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_upload.R
\name{mod_specid_batch_upload_server}
\alias{mod_specid_batch_upload_server}
\title{Batch upload server}
\usage{
mod_specid_batch_upload_server(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Reactive translator}
}
\value{
Reactive returning the loaded data frame (or NULL)
}
\description{
Batch upload server
}
\keyword{internal}
