% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_upload.R
\name{mod_specid_batch_upload_ui}
\alias{mod_specid_batch_upload_ui}
\title{Batch upload UI}
\usage{
mod_specid_batch_upload_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Batch upload UI
}
\keyword{internal}
