% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_validation.R
\name{mod_specid_batch_validation_server}
\alias{mod_specid_batch_validation_server}
\title{Batch validation server}
\usage{
mod_specid_batch_validation_server(
  id,
  matched_data,
  mappings,
  pool_main,
  pool_taxa,
  i18n
)
}
\arguments{
\item{id}{Module id}

\item{matched_data}{Reactive matched data (with id_colnam if applicable)}

\item{mappings}{Reactive mapping list}

\item{pool_main}{Reactive main DB pool}

\item{pool_taxa}{Reactive taxa DB pool}

\item{i18n}{Reactive translator}
}
\value{
list(validated_data, is_valid)
}
\description{
Batch validation server
}
\keyword{internal}
