% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_batch_validation.R
\name{mod_specid_batch_validation_ui}
\alias{mod_specid_batch_validation_ui}
\title{Batch validation UI}
\usage{
mod_specid_batch_validation_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Batch validation UI
}
\keyword{internal}
