% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_manual.R
\name{mod_specid_manual_server}
\alias{mod_specid_manual_server}
\title{Manual mode server}
\usage{
mod_specid_manual_server(
  id,
  pool_main,
  pool_taxa,
  i18n,
  active = shiny::reactive(TRUE)
)
}
\arguments{
\item{id}{Module id}

\item{pool_main}{Reactive main DB pool}

\item{pool_taxa}{Reactive taxa DB pool}

\item{i18n}{Reactive translator}

\item{active}{Reactive logical, TRUE when this mode is active}
}
\description{
Manual mode server
}
\keyword{internal}
