% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_manual.R
\name{mod_specid_manual_ui}
\alias{mod_specid_manual_ui}
\title{Manual mode UI}
\usage{
mod_specid_manual_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Manual mode UI
}
\keyword{internal}
