% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_mode.R
\name{mod_specid_mode_server}
\alias{mod_specid_mode_server}
\title{Mode selector server}
\usage{
mod_specid_mode_server(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Reactive translator}
}
\value{
Reactive returning "manual" or "batch" (NULL until picked)
}
\description{
Mode selector server
}
\keyword{internal}
