% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specid_mode.R
\name{mod_specid_mode_ui}
\alias{mod_specid_mode_ui}
\title{Mode selector UI}
\usage{
mod_specid_mode_ui(id, i18n)
}
\arguments{
\item{id}{Module id}

\item{i18n}{Translator}
}
\description{
Mode selector UI
}
\keyword{internal}
