% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_add.R
\name{mod_specimen_add_server}
\alias{mod_specimen_add_server}
\title{Specimen Add Module - Server}
\usage{
mod_specimen_add_server(id, pool_main, pool_taxa, i18n)
}
\arguments{
\item{id}{Character, module namespace ID}

\item{pool_main}{Reactive returning database connection pool}

\item{pool_taxa}{Reactive returning taxa database connection pool}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive with newly added specimen info
}
\description{
Specimen Add Module - Server
}
