% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_import.R
\name{mod_specimen_import_server}
\alias{mod_specimen_import_server}
\title{Specimen Import Module - Server}
\usage{
mod_specimen_import_server(
  id,
  matched_data,
  mappings,
  matching_complete,
  con,
  i18n
)
}
\arguments{
\item{id}{Module namespace ID}

\item{matched_data}{Reactive containing matched data with IDs}

\item{mappings}{Reactive containing column mappings}

\item{matching_complete}{Reactive boolean indicating if matching is complete}

\item{con}{Reactive database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing import results
}
\description{
Specimen Import Module - Server
}
\keyword{internal}
