% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_lookup.R
\name{mod_specimen_lookup_server}
\alias{mod_specimen_lookup_server}
\title{Specimen Lookup Module - Server}
\usage{
mod_specimen_lookup_server(id, data, mappings, con_main, con_taxa, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive containing uploaded user data}

\item{mappings}{Reactive containing column mappings}

\item{con_main}{Reactive main database connection pool}

\item{con_taxa}{Reactive taxa database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing matched data and status
}
\description{
Specimen Lookup Module - Server
}
\keyword{internal}
