% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_mapping.R
\name{mod_specimen_mapping_server}
\alias{mod_specimen_mapping_server}
\title{Specimen Mapping Module - Server}
\usage{
mod_specimen_mapping_server(id, data, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive containing uploaded user data}

\item{con}{Reactive database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing column mappings
}
\description{
Specimen Mapping Module - Server
}
\keyword{internal}
