% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_mapping.R
\name{mod_specimen_mapping_ui}
\alias{mod_specimen_mapping_ui}
\title{Specimen Mapping Module - UI}
\usage{
mod_specimen_mapping_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Specimen Mapping Module - UI
}
\keyword{internal}
