% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_retriever.R
\name{mod_specimen_retriever_server}
\alias{mod_specimen_retriever_server}
\title{Specimen Retriever Module - Server}
\usage{
mod_specimen_retriever_server(id, parsed_data, collector_matches, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{parsed_data}{Reactive containing parsed herbarium data from Step 2}

\item{collector_matches}{Reactive containing collector ID matches from Step 3}

\item{con}{Reactive database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing retrieved specimens
}
\description{
Specimen Retriever Module - Server
}
\keyword{internal}
