% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_search.R
\name{mod_specimen_search_server}
\alias{mod_specimen_search_server}
\title{Specimen Search Module - Server}
\usage{
mod_specimen_search_server(id, pool_main, pool_taxa, i18n)
}
\arguments{
\item{id}{Character, module namespace ID}

\item{pool_main}{Reactive returning database connection pool}

\item{pool_taxa}{Reactive returning taxa database connection pool}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive list with selected specimens
}
\description{
Specimen Search Module - Server
}
