% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_search.R
\name{mod_specimen_search_ui}
\alias{mod_specimen_search_ui}
\title{Specimen Search Module - UI}
\usage{
mod_specimen_search_ui(id, i18n)
}
\arguments{
\item{id}{Character, module namespace ID}

\item{i18n}{shiny.i18n translator object}
}
\value{
Shiny UI element
}
\description{
Specimen Search Module - UI
}
