% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_upload.R
\name{mod_specimen_upload_server}
\alias{mod_specimen_upload_server}
\title{Specimen Upload Module - Server}
\usage{
mod_specimen_upload_server(id, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{con}{Reactive database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive value containing uploaded data (data frame)
}
\description{
Specimen Upload Module - Server
}
\keyword{internal}
