% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_specimen_upload.R
\name{mod_specimen_upload_ui}
\alias{mod_specimen_upload_ui}
\title{Specimen Upload Module - UI}
\usage{
mod_specimen_upload_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Specimen Upload Module - UI
}
\keyword{internal}
