% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step1_choose_type.R
\name{mod_step1_choose_type_server}
\alias{mod_step1_choose_type_server}
\title{Step 1 Module: Choose Import Type - Server}
\usage{
mod_step1_choose_type_server(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\value{
Reactive value containing selected import type ("plots" or "individuals")
}
\description{
Step 1 Module: Choose Import Type - Server
}
\keyword{internal}
