% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step1_choose_type.R
\name{mod_step1_choose_type_ui}
\alias{mod_step1_choose_type_ui}
\title{Step 1 Module: Choose Import Type - UI}
\usage{
mod_step1_choose_type_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 1 Module: Choose Import Type - UI
}
\keyword{internal}
