% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_upload.R
\name{mod_step2_upload_server}
\alias{mod_step2_upload_server}
\title{Step 2 Module: Upload Data - Server}
\usage{
mod_step2_upload_server(id, import_type, config, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{import_type}{Reactive value containing import type ("plots" or "individuals")}

\item{config}{Reactive value containing import configuration}

\item{con}{Reactive database connection pool}

\item{i18n}{Translator object from shiny.i18n}
}
\value{
Reactive value containing uploaded data (data frame)
}
\description{
Step 2 Module: Upload Data - Server
}
\keyword{internal}
