% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step2_upload.R
\name{mod_step2_upload_ui}
\alias{mod_step2_upload_ui}
\title{Step 2 Module: Upload Data - UI}
\usage{
mod_step2_upload_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 2 Module: Upload Data - UI
}
\keyword{internal}
