% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step3_mapping.R
\name{mod_step3_mapping_ui}
\alias{mod_step3_mapping_ui}
\title{Step 3 Module: Column Mapping - UI}
\usage{
mod_step3_mapping_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 3 Module: Column Mapping - UI
}
\keyword{internal}
