% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step4_lookup_matching.R
\name{mod_step4_lookup_matching_ui}
\alias{mod_step4_lookup_matching_ui}
\title{Step 4 Module: Lookup Matching - UI}
\usage{
mod_step4_lookup_matching_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 4 Module: Lookup Matching - UI
}
\keyword{internal}
