% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step5_validation.R
\name{mod_step5_validation_server}
\alias{mod_step5_validation_server}
\title{Step 5 Module: Data Validation - Server}
\usage{
mod_step5_validation_server(id, data, mappings, config, con)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive containing uploaded user data (should be matched data from Step 4)}

\item{mappings}{Reactive containing column mappings}

\item{config}{Reactive containing import configuration}

\item{con}{Reactive containing database connection pool}
}
\value{
Reactive containing validation results
}
\description{
Step 5 Module: Data Validation - Server
}
\keyword{internal}
