% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step5_validation.R
\name{mod_step5_validation_ui}
\alias{mod_step5_validation_ui}
\title{Step 5 Module: Data Validation - UI}
\usage{
mod_step5_validation_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 5 Module: Data Validation - UI
}
\keyword{internal}
