% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step6_preview.R
\name{mod_step6_preview_server}
\alias{mod_step6_preview_server}
\title{Step 6 Module: Preview Data - Server}
\usage{
mod_step6_preview_server(id, validation_result)
}
\arguments{
\item{id}{Module namespace ID}

\item{validation_result}{Reactive containing validation results}
}
\value{
Reactive indicating preview confirmed
}
\description{
Step 6 Module: Preview Data - Server
}
\keyword{internal}
