% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step6_preview.R
\name{mod_step6_preview_ui}
\alias{mod_step6_preview_ui}
\title{Step 6 Module: Preview Data - UI}
\usage{
mod_step6_preview_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 6 Module: Preview Data - UI
}
\keyword{internal}
