% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step7_import.R
\name{mod_step7_import_ui}
\alias{mod_step7_import_ui}
\title{Step 7 Module: Execute Import - UI}
\usage{
mod_step7_import_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Step 7 Module: Execute Import - UI
}
\keyword{internal}
