% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_add.R
\name{mod_taxa_add_ui}
\alias{mod_taxa_add_ui}
\title{Taxa Add Module - UI}
\usage{
mod_taxa_add_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for adding new taxonomic entries
}
\keyword{internal}
