% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_r_code.R
\name{mod_taxa_r_code_ui}
\alias{mod_taxa_r_code_ui}
\title{Taxa R Code Preview Module - UI}
\usage{
mod_taxa_r_code_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying equivalent R code for taxa search and trait
extraction. Shown/hidden by a toggle button.
}
\keyword{internal}
