% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_search.R
\name{mod_taxa_search_ui}
\alias{mod_taxa_search_ui}
\title{Taxa Search & Browser Module - UI}
\usage{
mod_taxa_search_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for searching and browsing taxonomic records
}
\keyword{internal}
