% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_synonymy.R
\name{mod_taxa_synonymy_ui}
\alias{mod_taxa_synonymy_ui}
\title{Taxa Synonymy Module - UI}
\usage{
mod_taxa_synonymy_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for managing taxonomic synonymy
}
\keyword{internal}
