% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_traits_table.R
\name{mod_taxa_traits_table_ui}
\alias{mod_taxa_traits_table_ui}
\title{Taxa Traits Table Module - UI}
\usage{
mod_taxa_traits_table_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
Displays taxa-level traits as formatted wide/long tables with citation info,
for a single selected taxon. Used inside mod_taxa_search.
}
\keyword{internal}
