% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_tree_view.R
\name{mod_taxa_tree_view_server}
\alias{mod_taxa_tree_view_server}
\title{Taxa Tree View Module - Server}
\usage{
mod_taxa_tree_view_server(id, pool, selected_taxon, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{pool}{Reactive returning taxa database connection pool}

\item{selected_taxon}{Reactive returning selected taxon data from search module}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\description{
Server logic for displaying taxonomic hierarchy
}
\keyword{internal}
