% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_tree_view.R
\name{mod_taxa_tree_view_ui}
\alias{mod_taxa_tree_view_ui}
\title{Taxa Tree View Module - UI}
\usage{
mod_taxa_tree_view_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying taxonomic hierarchy as a tree view
}
\keyword{internal}
