% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_update.R
\name{mod_taxa_update_server}
\alias{mod_taxa_update_server}
\title{Taxa Update Module - Server}
\usage{
mod_taxa_update_server(id, pool, selected_taxon, has_write_permission, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{pool}{Reactive returning taxa database connection pool}

\item{selected_taxon}{Reactive returning selected taxon data from search module}

\item{has_write_permission}{Reactive returning TRUE if user can write}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\description{
Server logic for updating taxonomic records
}
\keyword{internal}
