% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_update.R
\name{mod_taxa_update_ui}
\alias{mod_taxa_update_ui}
\title{Taxa Update Module - UI}
\usage{
mod_taxa_update_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for updating existing taxonomic records
}
\keyword{internal}
