% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxo_match_r_code.R
\name{mod_taxo_match_r_code_ui}
\alias{mod_taxo_match_r_code_ui}
\title{Taxonomic Matching R Code Preview Module - UI}
\usage{
mod_taxo_match_r_code_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for displaying the equivalent R code of the automatic
matching. Shown/hidden by a toggle button.
}
\keyword{internal}
