% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomic_validator.R
\name{mod_taxonomic_validator_server}
\alias{mod_taxonomic_validator_server}
\title{Taxonomic Validator Module - Server}
\usage{
mod_taxonomic_validator_server(id, preliminary_links, con_taxa, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{preliminary_links}{Reactive containing retrieved specimens from Step 4}

\item{con_taxa}{Reactive taxa database connection pool}

\item{i18n}{Reactive returning translator object}
}
\value{
Reactive list containing validated links
}
\description{
Taxonomic Validator Module - Server
}
\keyword{internal}
