% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxonomic_validator.R
\name{mod_taxonomic_validator_ui}
\alias{mod_taxonomic_validator_ui}
\title{Taxonomic Validator Module - UI}
\usage{
mod_taxonomic_validator_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Taxonomic Validator Module - UI
}
\keyword{internal}
