% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_citation.R
\name{mod_trait_citation_ui}
\alias{mod_trait_citation_ui}
\title{Trait Citation Module - UI}
\usage{
mod_trait_citation_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Trait Citation Module - UI
}
\keyword{internal}
