% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_column_mapping.R
\name{mod_trait_column_mapping_ui}
\alias{mod_trait_column_mapping_ui}
\title{Trait Column Mapping Module - UI}
\usage{
mod_trait_column_mapping_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Trait Column Mapping Module - UI
}
\keyword{internal}
