% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_metadata_mapping.R
\name{mod_trait_metadata_mapping_server}
\alias{mod_trait_metadata_mapping_server}
\title{Trait Metadata Mapping Module - Server}
\usage{
mod_trait_metadata_mapping_server(id, data, trait_mapping, pool, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive: uploaded data frame}

\item{trait_mapping}{Reactive: result from mod_trait_column_mapping_server}

\item{pool}{Reactive: database connection pool}

\item{i18n}{Reactive: shiny.i18n translator}
}
\value{
Reactive list: valid, idtax_col, metadata_cols, feature_cols, available_traits
}
\description{
Trait Metadata Mapping Module - Server
}
\keyword{internal}
