% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_metadata_mapping.R
\name{mod_trait_metadata_mapping_ui}
\alias{mod_trait_metadata_mapping_ui}
\title{Trait Metadata Mapping Module - UI}
\usage{
mod_trait_metadata_mapping_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Trait Metadata Mapping Module - UI
}
\keyword{internal}
