% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_preview_import.R
\name{mod_trait_preview_import_ui}
\alias{mod_trait_preview_import_ui}
\title{Trait Preview & Import Module - UI}
\usage{
mod_trait_preview_import_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Trait Preview & Import Module - UI
}
\keyword{internal}
