% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_validation.R
\name{mod_trait_validation_ui}
\alias{mod_trait_validation_ui}
\title{Trait Validation Module - UI}
\usage{
mod_trait_validation_ui(id, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{i18n}{Translator object from shiny.i18n}
}
\description{
Trait Validation Module - UI
}
\keyword{internal}
