% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_traits_enrichment.R
\name{mod_traits_enrichment_server}
\alias{mod_traits_enrichment_server}
\title{Traits Enrichment Module - Server}
\usage{
mod_traits_enrichment_server(id, results, column_name, i18n)
}
\arguments{
\item{id}{Character, module ID}

\item{results}{Reactive list from review module (contains final matched data)}

\item{column_name}{Reactive returning the selected column name containing taxa}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\description{
Traits Enrichment Module - Server
}
\keyword{internal}
