% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_record.R
\name{mod_update_record_server}
\alias{mod_update_record_server}
\title{Record update module - server}
\usage{
mod_update_record_server(
  id,
  entity = c("plot", "individual"),
  pool_main,
  pool_taxa,
  i18n
)
}
\arguments{
\item{id}{Character, module namespace id.}

\item{entity}{Either `"plot"` or `"individual"`.}

\item{pool_main}{Reactive returning the main database pool.}

\item{pool_taxa}{Reactive returning the taxa database pool.}

\item{i18n}{Reactive returning a `shiny.i18n` translator.}
}
\value{
Invisibly `NULL`.
}
\description{
Record update module - server
}
