% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_update_record.R
\name{mod_update_record_ui}
\alias{mod_update_record_ui}
\title{Record update module - UI}
\usage{
mod_update_record_ui(id, entity = c("plot", "individual"), i18n)
}
\arguments{
\item{id}{Character, module namespace id.}

\item{entity}{Either `"plot"` or `"individual"`.}

\item{i18n}{A `shiny.i18n` translator object.}
}
\value{
A Shiny UI element.
}
\description{
Record update module - UI
}
