% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{PlotFetcher}
\alias{PlotFetcher}
\title{Fetch plot data}
\description{
Class for extracting metadata from database
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PlotFetcher-initialize}{\code{PlotFetcher$new()}}
    \item \href{#method-PlotFetcher-fetch_by_ids}{\code{PlotFetcher$fetch_by_ids()}}
    \item \href{#method-PlotFetcher-fetch_with_filter}{\code{PlotFetcher$fetch_with_filter()}}
    \item \href{#method-PlotFetcher-clone}{\code{PlotFetcher$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotFetcher-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PlotFetcher-initialize}{}}}
\subsection{\code{PlotFetcher$new()}}{
  Initialiser le fetcher
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotFetcher$new(connection)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{connection}}{Connexion DBI}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotFetcher-fetch_by_ids"></a>}}
\if{latex}{\out{\hypertarget{method-PlotFetcher-fetch_by_ids}{}}}
\subsection{\code{PlotFetcher$fetch_by_ids()}}{
  Récupérer plots par IDs
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotFetcher$fetch_by_ids(plot_ids)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plot_ids}}{Vecteur d'IDs de plots}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotFetcher-fetch_with_filter"></a>}}
\if{latex}{\out{\hypertarget{method-PlotFetcher-fetch_with_filter}{}}}
\subsection{\code{PlotFetcher$fetch_with_filter()}}{
  Récupérer plots avec filtre (requête SQL)
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotFetcher$fetch_with_filter(query)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{query}}{Requête SQL construite par PlotFilterBuilder}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotFetcher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotFetcher-clone}{}}}
\subsection{\code{PlotFetcher$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PlotFetcher$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
