% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_documentation.R
\name{print.column_documentation}
\alias{print.column_documentation}
\title{Print method for column_documentation objects}
\usage{
\method{print}{column_documentation}(x, ...)
}
\arguments{
\item{x}{A column_documentation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for column_documentation objects
}
