% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{print_user_access_summary}
\alias{print_user_access_summary}
\title{Print user access summary}
\usage{
print_user_access_summary(con = NULL)
}
\arguments{
\item{con}{Connection to main database (optional).}
}
\description{
Prints a formatted summary of user access to the database.
}
