% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{proj_rel_xy}
\alias{proj_rel_xy}
\title{Project stems in geographical space}
\usage{
proj_rel_xy(coord_sf, coord_rel)
}
\arguments{
\item{coord_sf}{sf polygon output of query_plots, using extract_coordinates = TRUE}

\item{coord_rel}{tibble extract of query_plots individuals with relative coordinates}
}
\description{
Project stems in geogaphical space
}
\details{
The coord_rel should have the columns x_100 and y_100 that are the relative coordinates in the plot
}
\author{
Gilles Dauby
}
