% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen_linking_functions.R
\name{query_link_individual_specimen}
\alias{query_link_individual_specimen}
\title{Query Link Between Specimens and Individuals}
\usage{
query_link_individual_specimen(id_ind = NULL, id_specimen = NULL, con = NULL)
}
\arguments{
\item{id_ind}{Integer vector of individual IDs to query}

\item{id_specimen}{Integer vector of specimen IDs to query}

\item{con}{Database connection. If NULL, calls call.mydb()}
}
\value{
Tibble with link information including link type
}
\description{
Query existing links between specimens and individuals. Returns link type
information when available.
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
