% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{query_traits_measures}
\alias{query_traits_measures}
\title{Legacy function - wrapper for backward compatibility}
\usage{
query_traits_measures(
  idtax = NULL,
  idtax_good = NULL,
  add_taxa_info = FALSE,
  id_trait = NULL,
  trait_cat_mode = "most_frequent",
  verbose = TRUE,
  pivot_table = TRUE,
  include_remarks = FALSE,
  extract_trait_measures_features = FALSE,
  con = NULL,
  con_taxa = NULL
)
}
\description{
Legacy function - wrapper for backward compatibility
}
\keyword{internal}
