% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{SpecimenFetcher}
\alias{SpecimenFetcher}
\title{Specimen Fetcher}
\description{
R6 class for fetching specimen data from database
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SpecimenFetcher-initialize}{\code{SpecimenFetcher$new()}}
    \item \href{#method-SpecimenFetcher-fetch_by_ids}{\code{SpecimenFetcher$fetch_by_ids()}}
    \item \href{#method-SpecimenFetcher-fetch_by_collector_and_number}{\code{SpecimenFetcher$fetch_by_collector_and_number()}}
    \item \href{#method-SpecimenFetcher-fetch_with_filter}{\code{SpecimenFetcher$fetch_with_filter()}}
    \item \href{#method-SpecimenFetcher-clone}{\code{SpecimenFetcher$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFetcher-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFetcher-initialize}{}}}
\subsection{\code{SpecimenFetcher$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFetcher$new(connection)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFetcher-fetch_by_ids"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFetcher-fetch_by_ids}{}}}
\subsection{\code{SpecimenFetcher$fetch_by_ids()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFetcher$fetch_by_ids(specimen_ids)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFetcher-fetch_by_collector_and_number"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFetcher-fetch_by_collector_and_number}{}}}
\subsection{\code{SpecimenFetcher$fetch_by_collector_and_number()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFetcher$fetch_by_collector_and_number(id_table_colnam, colnbr)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFetcher-fetch_with_filter"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFetcher-fetch_with_filter}{}}}
\subsection{\code{SpecimenFetcher$fetch_with_filter()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFetcher$fetch_with_filter(query)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFetcher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFetcher-clone}{}}}
\subsection{\code{SpecimenFetcher$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFetcher$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
