% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{traits_taxa_list}
\alias{traits_taxa_list}
\title{List of trait}
\usage{
traits_taxa_list(id_trait = NULL, con = NULL)
}
\value{
A tibble of all traits
}
\description{
Provide list of traits available
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
