% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{update_dico_name}
\alias{update_dico_name}
\title{Update taxonomic data}
\usage{
update_dico_name(
  genus_searched = NULL,
  tax_esp_searched = NULL,
  tax_fam_searched = NULL,
  tax_order_searched = NULL,
  id_searched = NULL,
  new_tax_gen = NULL,
  new_tax_esp = NULL,
  new_tax_fam = NULL,
  new_tax_order = NULL,
  new_tax_rank1 = NULL,
  new_tax_rank = NULL,
  new_tax_name1 = NULL,
  new_tax_famclass = NULL,
  new_introduced_status = NULL,
  new_tax_rankesp = NULL,
  new_author1 = NULL,
  new_author2 = NULL,
  new_author3 = NULL,
  ask_before_update = TRUE,
  add_backup = TRUE,
  show_results = TRUE,
  cancel_synonymy = FALSE,
  synonym_of = NULL,
  exact_match = FALSE,
  con = NULL
)
}
\arguments{
\item{genus_searched}{string genus name searched}

\item{tax_esp_searched}{string species name searched}

\item{tax_fam_searched}{string family name searched}

\item{new_tax_gen}{string new genus name}

\item{new_tax_esp}{string new species name}

\item{new_tax_fam}{string new family name}

\item{new_tax_rank1}{string new rank}

\item{new_tax_name1}{string new name of rank1}

\item{new_author1}{string new author of the species name}

\item{new_author2}{string new author of the infraspecific name}

\item{new_author3}{string new additional author}

\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup of modified data}

\item{show_results}{logical TRUE by default, show the data that has been modified}

\item{synonym_of}{list if the selected taxa should be put in synonymy with an existing taxa, add in a list at least one values to identify to which taxa it will be put in synonymy: genus, species or id}

\item{new_full_name_auth}{string new full name with authors}

\item{new_taxook}{integer new tax code}

\item{new_morphocat}{integer new morphocat code}

\item{new_detvalue}{integer new detvalue code}

\item{new_full_name_no_auth}{string new full name without authors - if not provided concatenate of new_esp and new_genus}

\item{new_full_name_used}{string new full_name_used}

\item{new_full_name_used2}{string new full_name_used2}

\item{id_search}{integer id of the taxa searched}

\item{no_synonym_modif}{logical FALSE by default, if TRUE and if the selected taxa is considered as synonym, then this will be modified and the selected taxa will not longer be a synonym}
}
\value{
No return value individuals updated
}
\description{
Update taxonomic data
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
