% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{update_specimens_batch}
\alias{update_specimens_batch}
\title{Update specimens data data}
\usage{
update_specimens_batch(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  id_col = 1,
  launch_update = FALSE,
  add_backup = TRUE
)
}
\arguments{
\item{new_data}{data frame data containing id and new values}

\item{col_names_select}{string plot name of the selected plots}

\item{col_names_corresp}{string of the selected plots}

\item{id_col}{integer indicate which name of col_names_select is the id for matching data}

\item{launch_update}{logical if TRUE updates are performed}

\item{add_backup}{logical whether backup of modified data should be recorded}
}
\value{
No return value individuals updated
}
\description{
`r lifecycle::badge("deprecated")`

Superseded by \code{\link{update_records}} with
\code{table_type = "specimens"} and \code{method = "batch"}. Rename the
columns of \code{new_data} to their database names beforehand (the
\code{col_names_select} / \code{col_names_corresp} pair has no equivalent),
then:

\preformatted{
update_records(data = new_data, table_type = "specimens",
               method = "batch", execute = TRUE)
}

Update specimens data plot _ at a time
}
\seealso{
\code{\link{update_records}}
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
\keyword{internal}
