% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_update_functions_old.R
\name{add_entry_taxa}
\alias{add_entry_taxa}
\title{Add new entry to taxonomic table}
\usage{
add_entry_taxa(
  search_name_tps = NULL,
  tax_gen = NULL,
  tax_esp = NULL,
  tax_fam = NULL,
  tax_order = NULL,
  tax_famclass = NULL,
  tax_rank1 = NULL,
  tax_name1 = NULL,
  tax_rank2 = NULL,
  tax_name2 = NULL,
  author1 = NULL,
  author2 = NULL,
  author3 = NULL,
  year_description = NULL,
  synonym_of = NULL,
  morpho_species = FALSE,
  TPS_KEY = "15ad0b4c-f0d3-46ab-b649-178f2c75724f",
  tax_tax = NULL
)
}
\arguments{
\item{tax_gen}{string genus name}

\item{tax_esp}{string species name}

\item{tax_fam}{string family name}

\item{tax_rank1}{string tax_rank1 name}

\item{tax_name1}{string tax_name1 name}

\item{synonym_of}{list if the new entry should be put in synonymy with an existing taxa, add in a list at least one values to identify to which taxa it will be put in synonymy: genus, species or id}

\item{detvalue}{integer detvalue code}

\item{morphocat}{integer morphocat code}

\item{full_name}{string full name : genus + species + authors}
}
\value{
A tibble
}
\description{
Add new entry to taxonomic table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
