% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_trait}
\alias{add_trait}
\title{Add trait}
\usage{
add_trait(
  new_trait = NULL,
  new_relatedterm = NULL,
  new_valuetype = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_traitdescription = NULL,
  new_factorlevels = NULL,
  new_expectedunit = NULL,
  new_comments = NULL,
  new_category = NULL,
  con = NULL,
  interactive = TRUE
)
}
\arguments{
\item{new_trait}{A single string.}

\item{new_relatedterm}{Optional. A single string.}

\item{new_valuetype}{A single string, one of `"numeric"`, `"integer"`, `"categorical"`, `"ordinal"`, `"logical"`, `"character"`, `"table_data_liste_plots"`, or `"table_colnam"`.}

\item{new_maxallowedvalue}{Optional. if valuetype is numeric, indicate the maximum allowed value}

\item{new_minallowedvalue}{Optional. if valuetype is numeric, indicate the minimum allowed value}

\item{new_traitdescription}{Optional. A single string.}

\item{new_factorlevels}{Optional. Factor levels.}

\item{new_expectedunit}{Optional. A single string.}

\item{new_comments}{Optional. A single string.}
}
\value{
The function writes to a database table if confirmed by the user. The function
will error if `new_trait` or `new_valuetype` are not provided, if `new_valuetype`
is not one of the allowed values, or if numeric/integer value types don't match
their corresponding min/max values.
}
\description{
Add trait and associated descriptors in trait list table
}
