% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_update_functions.R
\name{.add_taxa_noninteractive}
\alias{.add_taxa_noninteractive}
\title{Add taxonomic entry non-interactively (for Shiny apps)}
\usage{
.add_taxa_noninteractive(
  tax_gen = NULL,
  tax_esp = NULL,
  tax_fam = NULL,
  tax_order = NULL,
  tax_famclass = NULL,
  tax_rank1 = NULL,
  tax_name1 = NULL,
  author1 = NULL,
  author2 = NULL,
  author3 = NULL,
  year_description = NULL,
  morpho_species = FALSE,
  tax_tax = NULL,
  con = NULL
)
}
\arguments{
\item{tax_gen}{Character, genus name}

\item{tax_esp}{Character, species epithet}

\item{tax_fam}{Character, family name}

\item{tax_order}{Character, order name}

\item{tax_famclass}{Character, class name}

\item{tax_rank1}{Character, infraspecific rank ("var." or "subsp.")}

\item{tax_name1}{Character, infraspecific name}

\item{author1}{Character, species author}

\item{author2}{Character, infraspecific author}

\item{author3}{Character, additional author}

\item{year_description}{Numeric, year of description}

\item{morpho_species}{Logical, whether this is a morphospecies}

\item{tax_tax}{Character, full taxonomic name with authors}

\item{con}{Database connection (pool or connection object)}
}
\value{
Integer, the new idtax_n value
}
\description{
Non-interactive version of add_entry_taxa for use in Shiny applications.
Does not prompt for user input or growth forms.
}
\keyword{internal}
