% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_measurements.R
\name{.apply_long_mapping}
\alias{.apply_long_mapping}
\title{Apply long format mapping and build prepared data}
\usage{
.apply_long_mapping(
  df,
  traits,
  census_map,
  input,
  ns,
  i18n,
  link_policy = character(0)
)
}
\description{
Apply long format mapping and build prepared data
}
\keyword{internal}
