% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feat_step3_measurements.R
\name{.apply_wide_mapping}
\alias{.apply_wide_mapping}
\title{Apply wide format mapping and build prepared data}
\usage{
.apply_wide_mapping(df, traits, census_map, input, ns, i18n)
}
\description{
Apply wide format mapping and build prepared data
}
\keyword{internal}
