% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations_standardization.R
\name{.fix_word_boundaries}
\alias{.fix_word_boundaries}
\title{Convert `\b` to unicode-safe (?<![[:alpha:]]) / (?![[:alpha:]])
depending on whether the boundary is at the start or end of a token.
Called from \code{.load_observations_ontology}.}
\usage{
.fix_word_boundaries(pattern)
}
\description{
Convert `\b` to unicode-safe (?<![[:alpha:]]) / (?![[:alpha:]])
depending on whether the boundary is at the start or end of a token.
Called from \code{.load_observations_ontology}.
}
\keyword{internal}
