% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{enrich_individual_traits}
\alias{enrich_individual_traits}
\title{Enrich with individual-level traits}
\usage{
enrich_individual_traits(
  individuals,
  con,
  show_multiple_census,
  issues = c("remove", "include", "ignore"),
  include_measurement_ids = FALSE,
  census_strategy = c("last", "first", "mean"),
  individual_features_format = c("wide", "long", "census_pairs")
)
}
\description{
Enrich with individual-level traits
}
\keyword{internal}
