% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{match_tax}
\alias{match_tax}
\title{Query and standardize taxonomy}
\usage{
match_tax(
  idtax,
  queried_tax = NULL,
  verbose = TRUE,
  backbone = c("internal", "wcvp")
)
}
\arguments{
\item{idtax}{vector of idtax_n to be search}

\item{queried_tax}{tibble, output of query_taxa}

\item{verbose}{logical whether results should be shown in viewer}

\item{backbone}{Character. Which taxonomic backbone to use for synonym resolution.
\code{"internal"} (default) uses the internal \code{table_taxa}.
\code{"wcvp"} uses WCVP via \code{wcvp_idtax_link} and \code{wcvp_names},
falling back to internal for unlinked taxa.}
}
\value{
tibble
}
\description{
Query and standardize taxonomy for synonymies and add traits information at species and genus levels
}
\examples{
\dontrun{
match_tax(idtax = c(3095, 219))
}

}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
