% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_login.R
\name{mod_database_login_ui}
\alias{mod_database_login_ui}
\title{Database Login Module - UI}
\usage{
mod_database_login_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A shiny tagList
}
\description{
UI component for database authentication with language selection
}
\keyword{internal}
