% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lookup_matcher.R
\name{mod_lookup_matcher_server}
\alias{mod_lookup_matcher_server}
\title{Lookup Matcher Module - Server}
\usage{
mod_lookup_matcher_server(id, invalid_values, con)
}
\arguments{
\item{id}{Module namespace ID}

\item{invalid_values}{Reactive list of invalid values by column (e.g., list(method = c("val1", "val2")))}

\item{con}{Reactive database connection pool}
}
\value{
Reactive list of resolved matches
}
\description{
Lookup Matcher Module - Server
}
\keyword{internal}
