% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_display.R
\name{mod_results_display_server}
\alias{mod_results_display_server}
\title{Results Display Module - Server}
\usage{
mod_results_display_server(id, results, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{results}{Reactive containing query_plots() results}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\description{
Server logic for results display and download
}
\keyword{internal}
