% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step3_mapping.R
\name{mod_step3_mapping_server}
\alias{mod_step3_mapping_server}
\title{Step 3 Module: Column Mapping - Server}
\usage{
mod_step3_mapping_server(id, data, config, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive containing uploaded user data}

\item{config}{Reactive containing import configuration}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Translator object from shiny.i18n}
}
\value{
Reactive list containing mappings and validation status
}
\description{
Step 3 Module: Column Mapping - Server
}
\keyword{internal}
