% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_step4_lookup_matching.R
\name{mod_step4_lookup_matching_server}
\alias{mod_step4_lookup_matching_server}
\title{Step 4 Module: Lookup Matching - Server}
\usage{
mod_step4_lookup_matching_server(id, data, mappings, config, con, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive containing uploaded user data}

\item{mappings}{Reactive containing column mappings}

\item{config}{Reactive containing import configuration}

\item{con}{Reactive containing database connection pool}

\item{i18n}{Reactive returning translator object from shiny.i18n}
}
\value{
Reactive containing matched/updated data
}
\description{
Step 4 Module: Lookup Matching - Server
}
\keyword{internal}
