% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_add.R
\name{mod_taxa_add_server}
\alias{mod_taxa_add_server}
\title{Taxa Add Module - Server}
\usage{
mod_taxa_add_server(id, pool, has_write_permission, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{pool}{Reactive returning taxa database connection pool}

\item{has_write_permission}{Reactive returning TRUE if user can write}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\description{
Server logic for adding new taxonomic entries
}
\keyword{internal}
