% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxa_search.R
\name{mod_taxa_search_server}
\alias{mod_taxa_search_server}
\title{Taxa Search & Browser Module - Server}
\usage{
mod_taxa_search_server(id, pool, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{pool}{Reactive returning taxa database connection pool}

\item{i18n}{Reactive returning shiny.i18n translator}
}
\value{
Reactive containing selected taxon data (full row from table_taxa)
}
\description{
Server logic for taxonomic search and browsing
}
\keyword{internal}
