% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trait_preview_import.R
\name{mod_trait_preview_import_server}
\alias{mod_trait_preview_import_server}
\title{Trait Preview & Import Module - Server}
\usage{
mod_trait_preview_import_server(id, data, mapping, pool, i18n)
}
\arguments{
\item{id}{Module namespace ID}

\item{data}{Reactive returning uploaded data frame}

\item{mapping}{Reactive returning mapping result from mod_trait_column_mapping_server}

\item{pool}{Reactive returning database connection pool}

\item{i18n}{Reactive returning translator}
}
\value{
Reactive list with import_result
}
\description{
Trait Preview & Import Module - Server
}
\keyword{internal}
