% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{query_subplots}
\alias{query_subplots}
\title{Legacy function - wrapper for backward compatibility}
\usage{
query_subplots(
  ids_plots = NULL,
  ids_subplots = NULL,
  plot_name = NULL,
  country = NULL,
  locality_name = NULL,
  method = NULL,
  subtype = NULL,
  verbose = TRUE,
  extract_subplots_obs_features = FALSE,
  con = NULL
)
}
\description{
Legacy function - wrapper for backward compatibility
}
\keyword{internal}
