% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_query_plots.R
\name{shiny_app_query_plots}
\alias{shiny_app_query_plots}
\title{Query Plots Shiny App}
\usage{
shiny_app_query_plots(pool_main = NULL, language = "fr")
}
\arguments{
\item{pool_main}{Main database connection pool (optional, will prompt for login)}

\item{language}{Character, initial language ("en" or "fr"), default: "fr"}
}
\value{
A Shiny app object
}
\description{
Main Shiny app function for interactive forest plot querying
}
\keyword{internal}
