% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_specimen_import_wizard.R
\name{shiny_app_specimen_import_wizard}
\alias{shiny_app_specimen_import_wizard}
\title{Specimen Import Wizard (Internal Function)}
\usage{
shiny_app_specimen_import_wizard(lang = "en")
}
\arguments{
\item{lang}{Character, initial language}
}
\value{
Shiny app object
}
\description{
Internal function that creates the Shiny app object without launching it.
Useful for testing or embedding in other applications.
}
\keyword{internal}
