% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{SpecimenFilterBuilder}
\alias{SpecimenFilterBuilder}
\title{Specimen Filter Builder}
\description{
R6 class for building SQL queries to filter specimens with a fluent API
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SpecimenFilterBuilder-initialize}{\code{SpecimenFilterBuilder$new()}}
    \item \href{#method-SpecimenFilterBuilder-filter_collector}{\code{SpecimenFilterBuilder$filter_collector()}}
    \item \href{#method-SpecimenFilterBuilder-filter_number}{\code{SpecimenFilterBuilder$filter_number()}}
    \item \href{#method-SpecimenFilterBuilder-filter_taxonomy}{\code{SpecimenFilterBuilder$filter_taxonomy()}}
    \item \href{#method-SpecimenFilterBuilder-filter_by_ids}{\code{SpecimenFilterBuilder$filter_by_ids()}}
    \item \href{#method-SpecimenFilterBuilder-build}{\code{SpecimenFilterBuilder$build()}}
    \item \href{#method-SpecimenFilterBuilder-add_custom_condition}{\code{SpecimenFilterBuilder$add_custom_condition()}}
    \item \href{#method-SpecimenFilterBuilder-print_conditions}{\code{SpecimenFilterBuilder$print_conditions()}}
    \item \href{#method-SpecimenFilterBuilder-clone}{\code{SpecimenFilterBuilder$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-initialize}{}}}
\subsection{\code{SpecimenFilterBuilder$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$new(connection)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-filter_collector"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-filter_collector}{}}}
\subsection{\code{SpecimenFilterBuilder$filter_collector()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$filter_collector(
  collector = NULL,
  id_colnam = NULL,
  interactive = FALSE
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-filter_number"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-filter_number}{}}}
\subsection{\code{SpecimenFilterBuilder$filter_number()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$filter_number(
  number = NULL,
  number_min = NULL,
  number_max = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-filter_taxonomy"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-filter_taxonomy}{}}}
\subsection{\code{SpecimenFilterBuilder$filter_taxonomy()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$filter_taxonomy(
  genus = NULL,
  species = NULL,
  family = NULL,
  idtax_n = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-filter_by_ids"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-filter_by_ids}{}}}
\subsection{\code{SpecimenFilterBuilder$filter_by_ids()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$filter_by_ids(specimen_ids)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-build"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-build}{}}}
\subsection{\code{SpecimenFilterBuilder$build()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$build(operator = "AND")}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-add_custom_condition"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-add_custom_condition}{}}}
\subsection{\code{SpecimenFilterBuilder$add_custom_condition()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$add_custom_condition(condition, wrap_parentheses = TRUE)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-print_conditions"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-print_conditions}{}}}
\subsection{\code{SpecimenFilterBuilder$print_conditions()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$print_conditions()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenFilterBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenFilterBuilder-clone}{}}}
\subsection{\code{SpecimenFilterBuilder$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenFilterBuilder$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
